+++ src/pages/prompt.vue
